sql: fix ORDER BY with aggregates.
Since 146344b799cb94f42789336bb156c614a06be9d5 we have this
optimization which elides a sort if the sort expression(s) are known
to be constant throughout all the rows being sorted.

However the test to decide whether to elide the sort was (incorrectly)
relying only on the presence of columns from the input data source;
which was later extended (but still incorrectly) to the "random"
function and "count" applied to a StarDatumInstance. This was still
incorrect because *any* aggregate function is susceptible to change
values between the rows being sorted -- the rows being sorted, with an
aggregate function, are the results of grouping, and the evaluated
value of the aggregate function call is different for every group.

This patch addresses the issue more durably by:

- introducing a new built-in attribute "RowDependent" which indicates
  whether a particular function is row-dependent. "random" and all
  aggregate functions are marked with this property.

- using this new property in the code that elides the sort.
