+++ pkg/sql/parser/function_name.go
@@ -19,7 +19,6 @@ package parser
 import (
 	"bytes"
 	"fmt"
-	"strings"
 )
 
 // Function names are used in expressions in the FuncExpr node.
@@ -138,75 +137,4 @@ func (fn *functionName) function() string {
 // Prefix retrieves the unqualified prefix.
 func (fn *functionName) prefix() string {
 	return string(fn.prefixName)
-}
-
-// FunctionDefinition implements a reference to one or more function
-// overloads for a built-in function.
-type FunctionDefinition struct {
-	Name       string
-	Definition []Builtin
-}
-
-// funDefs holds pre-allocated FunctionDefinition instances
-// for every builtin function. Initialized by setupBuiltins().
-var funDefs map[string]*FunctionDefinition
-
-// Format implements the NodeFormatter interface.
-func (fd *FunctionDefinition) Format(buf *bytes.Buffer, f FmtFlags) {
-	buf.WriteString(fd.Name)
-}
-
-func (fd *FunctionDefinition) String() string { return AsString(fd) }
-
-// SearchPath represents a list of namespaces to search builtins in.
-// The names must be normalized (as per Name.Normalize) already.
-type SearchPath []string
-
-// ResolveFunction transforms an UnresolvedName to a FunctionDefinition.
-func (n UnresolvedName) ResolveFunction(searchPath SearchPath) (*FunctionDefinition, error) {
-	fn, err := n.normalizeFunctionName()
-	if err != nil {
-		return nil, err
-	}
-
-	if len(fn.selector) > 0 {
-		// We do not support selectors at this point.
-		return nil, fmt.Errorf("invalid function name: %s", n)
-	}
-
-	if d, ok := funDefs[fn.function()]; ok && fn.prefix() == "" {
-		// Fast path: return early.
-		return d, nil
-	}
-
-	// Although the conversion from Name to string should go via
-	// Name.Normalize(), functions are special in that they are
-	// guaranteed to not contain special Unicode characters. So we can
-	// use ToLower directly.
-	prefix := strings.ToLower(fn.prefix())
-	smallName := strings.ToLower(fn.function())
-	fullName := smallName
-	if prefix != "" {
-		fullName = prefix + "." + smallName
-	}
-	def, ok := funDefs[fullName]
-	if !ok {
-		found := false
-		if prefix == "" {
-			// The function wasn't qualified, so we must search for it via
-			// the search path first.
-			for _, alt := range searchPath {
-				fullName = alt + "." + smallName
-				if def, ok = funDefs[fullName]; ok {
-					found = true
-					break
-				}
-			}
-		}
-		if !found {
-			return nil, fmt.Errorf("unknown function: %s()", n)
-		}
-	}
-
-	return def, nil
 }
